[#5] refactor: post 조회 로직 postCommand -> postQuery
