+++ gregorsamsa-api/src/main/kotlin/com/gregorsamsa/post/service/PostQueryService.kt
@@ -1,16 +1,17 @@
 package com.gregorsamsa.post.service
 
 import com.gregorsamsa.core.post.PostCommand
+import com.gregorsamsa.core.post.PostQuery
 import com.gregorsamsa.exception.NotFoundException
 import com.gregorsamsa.post.web.dto.response.PostViewResponse
 import org.springframework.stereotype.Service
 
 @Service
 class PostQueryService(
-    private val postCommand: PostCommand
+    private val postQuery: PostQuery
 ) {
     fun getAllPost(): List<PostViewResponse> {
-        val allPost = postCommand.getAllPost()
+        val allPost = postQuery.getAllPost()
 
         return allPost.map {
             PostViewResponse.of(it)
@@ -18,7 +19,7 @@ class PostQueryService(
     }
 
     fun getPost(postId: Long): PostViewResponse {
-        val post = postCommand.getPostOrNull(postId)
+        val post = postQuery.getPostOrNull(postId)
             ?: throw NotFoundException("post is null, postId:$postId")
 
         return PostViewResponse.of(post)
