+++ gregorsamsa-core/src/main/kotlin/com/gregorsamsa/core/post/PostCommand.kt
@@ -1,6 +1,5 @@
 package com.gregorsamsa.core.post
 
-import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDateTime
@@ -27,17 +26,5 @@ class PostCommand(
         val saved = postRepository.save(post)
 
         return saved.id!!
-    }
-
-    fun getAllPost(): List<PostVO> {
-        val posts = postRepository.findAll()
-
-        return posts.map { PostVO.of(it) }
-    }
-
-    fun getPostOrNull(postId: Long): PostVO? {
-        val post = postRepository.findByIdOrNull(postId)
-
-        return post?.let { PostVO.of(it) }
     }
 }
